Use long

+++ src/main/java/org/jenkinsci/plugins/github_branch_source/JwtHelper.java
@@ -42,7 +42,7 @@ class JwtHelper {
                 .setIssuer(githubAppId)
                 .signWith(signingKey, signatureAlgorithm);
 
-        int oneMinuteInMillis = 60 * 1000;
+        long oneMinuteInMillis = 60L * 1000L;
         long expMillis = nowMillis + (oneMinuteInMillis * 8);
         Date exp = new Date(expMillis);
         builder.setExpiration(exp);
